requirements.txt
discord.py==2.6.4
flask==2.2.3
werkzeug==2.2.3
hypercorn==0.15.0